+++ pkg/cmd/gobarista/commands/billing.go
@@ -264,6 +264,7 @@ var BillingAddBill = cli.Command{
 		if err != nil {
 			return fmt.Errorf("error: cannot add new bill to billing period: %v", err.Error())
 		}
+
 		logger.Info(fmt.Sprintf("new bill successfully added to billing period for user_id=%d, user_name=%s: new bill id: %d", uid, user.Firstname+" "+user.Lastname, id))
 
 		return nil
@@ -321,11 +322,76 @@ var BillingIssueBills = cli.Command{
 				logger.Error(fmt.Sprintf("error: billing e-mail has not been sent for bill_id=%d user_id=%d user_name='%s' user_email:'%s'", bill.ID, user.ID, user.Firstname+" "+user.Lastname, user.Email))
 			} else {
 				logger.Info(fmt.Sprintf("billing e-mail has been sent for bill_id=%d user_id=%d user_name='%s' user_email:'%s'", bill.ID, user.ID, user.Firstname+" "+user.Lastname, user.Email))
+
+				err = database.UpdateBillOnIssued(bill.ID)
+				if err != nil {
+					return fmt.Errorf("error: cannot update bill: bill_id=%d: %v", bill.ID, err)
+				}
+			}
+		}
+
+		return nil
+	},
+}
+
+var BillingConfirmPaymentBills = cli.Command{
+	Name:      "confirm-payment",
+	Aliases:   []string{"c"},
+	Usage:     "Send payment confirmation for all paid bills (e-mail will not be sent for already notified users)",
+	ArgsUsage: "[id]",
+	Action: func(ctx *cli.Context) (err error) {
+		if err = helpers.SetupDatabase(ctx); err != nil {
+			return err
+		}
+
+		if err = helpers.SetupMail(ctx); err != nil {
+			return err
+		}
+
+		if ctx.NArg() != 1 {
+			return fmt.Errorf("error: too few arguments: requires (1), get (%d)", ctx.NArg())
+		}
+
+		pid, err := strconv.ParseUint(ctx.Args().Get(0), 10, 64)
+		if err != nil {
+			return fmt.Errorf("error: cannot parse uint: %v", err)
+		}
+
+		period, err := database.SelectPeriodByID(uint(pid))
+		if errors.Is(err, gorm.ErrRecordNotFound) {
+			return fmt.Errorf("error: billing period not found")
+		} else if err != nil {
+			return fmt.Errorf("error: cannot get billing period: %v", err)
+		}
+
+		if !period.Closed {
+			return fmt.Errorf("error: cannot send payment confirmation: period is not closed")
+		}
+
+		bills, err := database.SelectAllBillsForPeriod(uint(pid))
+		if err != nil {
+			return fmt.Errorf("error: cannot get bills for given period: %v", err)
+		}
+
+		for _, bill := range bills {
+			user, err := database.SelectUserByID(bill.UserID)
+			if errors.Is(err, gorm.ErrRecordNotFound) {
+				return fmt.Errorf("error: user for giver bill is not found: user_id=%d", bill.UserID)
+			} else if err != nil {
+				return fmt.Errorf("error: cannot get user: user_id=%d: %v", bill.UserID, err)
 			}
 
-			err = database.UpdateBillOnIssued(bill.ID)
-			if err != nil {
-				return fmt.Errorf("error: cannot update bill: bill_id=%d: %v", bill.ID, err)
+			if !bill.PaymentConfirmation {
+				if err = mail.SendPaymentConfirmation(user, period, bill); err != nil {
+					logger.Error(fmt.Sprintf("error: billing e-mail has not been sent for bill_id=%d user_id=%d user_name='%s' user_email:'%s'", bill.ID, user.ID, user.Firstname+" "+user.Lastname, user.Email))
+				} else {
+					logger.Info(fmt.Sprintf("billing e-mail has been sent for bill_id=%d user_id=%d user_name='%s' user_email:'%s'", bill.ID, user.ID, user.Firstname+" "+user.Lastname, user.Email))
+
+					err = database.UpdateBillOnPaymentConfirmation(bill.ID)
+					if err != nil {
+						return fmt.Errorf("error: cannot update bill: bill_id=%d: %v", bill.ID, err)
+					}
+				}
 			}
 		}