+++ resources/embed.go
@@ -11,3 +11,5 @@ const DIR_TEMPLATES = "templates"
 // General constants
 const HTML_BILL_TEMPLATE = "bill.cs.html"
 const HTML_BILL_TEMPLATE_FULL_PATH = DIR_TEMPLATES + "/" + HTML_BILL_TEMPLATE
+const HTML_CONFIRM_TEMPLATE = "confirm.cs.html"
+const HTML_CONFIRM_TEMPLATE_FULL_PATH = DIR_TEMPLATES + "/" + HTML_CONFIRM_TEMPLATE